requirement.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.1.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
bleach==5.0.0
blinker==1.4
bokeh==3.0.2
Brlapi==0.8.4
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
cloudpickle==2.2.0
colorama==0.4.5
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu1
duplicity==0.8.22
entrypoints==0.4
fastapi==0.85.0
fasteners==0.17.3
fastjsonschema==2.15.1
fonttools==4.34.4
frozenlist==1.3.3
fs==2.4.16
future==0.18.2
gast==0.5.2
gym==0.26.2
gym-notices==0.0.8
gyp==0.1
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
httptools==0.5.0
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.15.1
ipyparallel==8.4.1
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.8.0
Jinja2==3.0.3
json5==0.9.10
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.11.1
jupyter-server==1.23.2
jupyter-server-proxy==3.2.2
jupyter-ydoc==0.2.2
jupyter_server_ydoc==0.2.0
jupyterlab==3.6.0a3
jupyterlab-nvdashboard==0.7.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
keyring==23.8.2
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.22.0
lxml==4.9.1
lz4==4.0.0+dfsg
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.2
nbclassic==0.4.8
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
notebook_shim==0.2.2
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.10.4
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.30
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
Pygments==2.12.0
PyGObject==3.42.2
PyJWT==2.4.0
PyNaCl==1.5.0
pynvml==11.4.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian==0.1.46
python-dotenv==0.21.0
python-multipart==0.0.5
pythran==0.11.0
pytz==2022.2.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
reportlab==3.6.11
requests==2.28.1
scipy==1.8.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
Send2Trash==1.8.1b0
simpervisor==0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2
ssh-import-id==5.11
starlette==0.20.4
supervisor==4.2.1
sympy==1.10.1
systemd-python==235
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
torch==1.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==15.0.0
urllib3==1.26.9
usb-creator==0.3.7
uvicorn==0.18.3
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.3
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
xyzservices==2022.9.0
y-py==0.5.4
yarl==1.8.1
ypy-websocket==0.3.2
zipp==1.0.0
